requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
cryptography==42.0.7
defusedxml==0.8.0rc2
distlib==0.3.8
Django==4.2.13
django-cors-headers==4.3.1
django-filter==24.2
django-phonenumber-field==7.3.0
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-spectacular==0.27.2
filelock==3.14.0
flake8==7.0.0
identify==2.5.36
idna==3.7
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
mccabe==0.7.0
nodeenv==1.9.0
oauthlib==3.2.2
phonenumbers==8.13.38
platformdirs==4.2.2
pre-commit==3.7.1
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.6.3
pydantic-settings==2.2.1
pydantic_core==2.16.3
pyflakes==3.2.0
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.2
requests-oauthlib==2.0.0
rpds-py==0.18.1
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
typing_extensions==4.12.0
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.26.2
